create missing recipient registry and filter out foreign events
